(datalake) fix column position datalake error
change final columns in all models' trips tables to be the same when output to the datalake to resolve data appending issues in datalake. Columns that do not exist for a model are made to be null. \nChanges to the write_to_datalake extension made so that column filter only called for data written to the datalake, not locally. Can be reverted later.