Updated and cleaned trip mode choice configs after initial calibration
Fixed the bug in escooter tour conditions to not allow other trip modes.
Removed the trip mode choice coefficients for drive transit, ride hail, ebike and escooter tour modes. These coefficients are not necessary because symmetry is assumed in these modes and this is defined through availability conditions in trip mode choice preprocessor.