modifying bike availability in resident trip mode choice #120
This fixes issue on having bike trips with 0 times. This issue happens due to school escorting model inserting stops on bike tours too far to be reached by bike. These trips are now accomplished by sr2 or sr3. sre2 and sr3 trips on bike tours are disadvantaged (utile -30) so they be picked as a last option